added the job class view

+++ SeaNam Job Card Controller/ui/jobcards/Job Card.xaml.cs
@@ -2,9 +2,6 @@
 using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Input;
-using FluentNHibernate.Conventions;
-using MaterialDesignThemes.Wpf;
 using SeaNam_Job_Card_Controller.model;
 using SeaNam_Job_Card_Controller.repo;
 using SeaNam_Job_Card_Controller.Utils;
@@ -58,23 +55,48 @@ namespace SeaNam_Job_Card_Controller.ui.jobcards
 
         private void AddWorkArea_OnClick(object sender, RoutedEventArgs e)
         {
-            var areaPage = new WorkAreaPage(WorkAreaList);
+            var areaPage = new WorkAreaPage(WorkAreaList, new WorkArea());
             areaPage.ShowDialog();
         }
 
         private void EditWorkArea_OnClick(object sender, RoutedEventArgs e)
         {
-            var area = (WorkArea)WorkAreaCombo.SelectedItem;
-            new WorkAreaPage(area).ShowDialog();
-            
+            var area = (WorkArea) WorkAreaCombo.SelectedItem;
+            if (area == null) return;
+            new WorkAreaPage(WorkAreaList, area).ShowDialog();
+            WorkAreaCombo.SelectedItem = area;
         }
 
         private void DeleteWorkArea_OnClick(object sender, RoutedEventArgs e)
         {
             var area = (WorkArea) WorkAreaCombo.SelectedItem;
+            if (area == null) return;
             _workAreaRepo.DeleteModel(area);
             WorkAreaList.Clear();
             WorkAreaList.AddAll(_workAreaRepo.LoadModels());
+        }
+
+        private void AddJobClass_OnClick(object sender, RoutedEventArgs e)
+        {
+            var areaPage = new JobClassPage(JobClassList, new JobClass());
+            areaPage.ShowDialog();
+        }
+
+        private void EditJobClass_OnClick(object sender, RoutedEventArgs e)
+        {
+            var area = JobClassCombo.SelectedItem as JobClass;
+            if (area == null) return;
+            new JobClassPage(JobClassList, area).ShowDialog();
+            JobClassCombo.SelectedItem = area;
+        }
+
+        private void DeleteJobClass_OnClick(object sender, RoutedEventArgs e)
+        {
+            var area = JobClassCombo.SelectedItem as JobClass;
+            if(area == null) return;
+            _classRepo.DeleteModel(area);
+            JobClassList.Clear();
+            JobClassList.AddAll(_classRepo.LoadModels());
         }
     }
 }
